+++ Samples/Frontline.GameDesign/Designs/兵种/DEquipGrade.cs
@@ -13,7 +13,7 @@ namespace Frontline.GameDesign
         public int max_level { get; set; }
         public JsonObject<int[]> grade_item_id { get; set; }
         public JsonObject<int[]> grade_item_cnt { get; set; }
-        public int grade_grow { get; set; }
+        public float grade_grow { get; set; }
         public int next_id { get; set; }
     }
 }
